:bricks: ci: style optimization
In this commit, several changes were made related to the structure and configuration of the site. First, we removed the global CSS import, which was no longer needed. Next, the import of the Inter font, which was not being used in the project, was removed. Then, changes were made to the site's title and description meta data in order to improve its relevance and optimize the search experience. Finally, the change of language from 'en' to 'pt-br' was made, to adapt the language of the site to the target audience and facilitate the understanding of visitors. These changes contribute to a cleaner structure, better performance and a better experience for website users.