:bricks: ci: usability improvement
In this commit, the default page was changed to a custom home page. The previous page, which was the automatically generated default, has been removed and replaced with a new home page. The new home page has an h1 element that contains the text 'Home', highlighting the purpose of the page.